Refactoring player info card
