+++ src/components/Navbar.vue
@@ -69,54 +69,7 @@
         </v-btn>
       </template>
 
-      <v-card :dark="darkMode">
-
-        <v-card-title :class="darkMode ? 'indigo darken-2' : 'indigo'">
-          <h1 class="text-h5 white--text card-title-shadow">Player Info</h1>
-        </v-card-title>
-
-        <v-card-text>
-
-          <v-row justify="center">
-            <v-avatar size=80 class="avatar-shadow">
-              <v-img :src="getAvatarPicture" alt="User"></v-img>
-            </v-avatar>
-          </v-row>
-
-          <v-row class="mt-5" justify="center">
-            <h1
-            :class="['text-h4', darkMode ? 'white--text' : 'black--text']"
-            >
-              {{user.username}}
-            </h1>
-          </v-row>
-
-          <v-row justify="center" class="mt-4">
-            <h1 class="text-subtitle-1 font-weight-regular">{{user.email}}</h1>
-          </v-row>
-
-          <v-row justify="center" class="mt-4">
-            <h1 class="text-subtitle-2 font-weight-regular">{{getAge(user.birthday)}} years old</h1>
-          </v-row>
-
-          <v-row justify="center" class="mt-6">
-            <h1
-            :class="['text-caption font-weight-medium',
-            darkMode ? 'white--text' : 'black--text']"
-            >
-              {{user.position}}
-            </h1>
-          </v-row>
-
-          <v-row justify="center" class="mt-5 mb-2">
-            <img
-            :class="[positionImage, 'playerImg']"
-            :src="require(`@/assets/myButtons/${positionImage}.png`)"
-            />
-          </v-row>
-
-        </v-card-text>
-      </v-card>
+      <player-info :darkMode="darkMode" :user="user" />
 
     </v-menu>
 
@@ -127,6 +80,7 @@
 import { mapGetters } from 'vuex';
 import BreakpointsCond from '../mixins/BreakpointsCond';
 import ModeSwitcher from './ModeSwitcher.vue';
+import PlayerInfo from './PlayerInfo.vue';
 /* eslint-disable global-require */
 
 export default {
@@ -134,6 +88,7 @@ export default {
 
   components: {
     ModeSwitcher,
+    PlayerInfo,
   },
 
   data() {
@@ -148,9 +103,6 @@ export default {
     getAvatarPicture() {
       // eslint-disable-next-line import/no-dynamic-require
       return require(`../${this.user.picture}`);
-    },
-    positionImage() {
-      return this.user.position.toLowerCase();
     },
   },
 
@@ -183,13 +135,6 @@ export default {
         this.$router.push({ name: 'Home' });
       }
     },
-
-    getAge(birthday) {
-      const bday = new Date(birthday);
-      const ageDiffMs = Date.now() - bday.getTime();
-      const ageDate = new Date(ageDiffMs);
-      return Math.abs(ageDate.getUTCFullYear() - 1970);
-    },
   },
 
   mixins: [BreakpointsCond],
@@ -214,30 +159,6 @@ export default {
 }
 .text-shadow {
   text-shadow: 1px 1px rgba(0, 0, 0, 0.4);
-}
-.card-title-shadow {
-  text-shadow: 2px 2px rgba(0, 0, 0, 0.7);
-}
-.playerImg {
-  /* centering img and default opacity*/
-  display: block;
-  margin-left: auto;
-  margin-right: auto;
-  filter: drop-shadow(2px 2px rgba(0, 0, 0, 0.6));
-}
-/* custom sizes and opacity based on type of button */
-.goalkeeper {
-  max-width: 58px;
-}
-.defender {
-  max-width: 28px;
-}
-.forward {
-  max-width: 70px;
-}
-.avatar-shadow {
-  filter: drop-shadow(0px 0px 5px rgba(0, 0, 0, 0.5));
-  border: 0.5px solid rgba(0, 0, 0, 0.2);
 }
 .avatar-shadow-nav {
   filter: drop-shadow(1px 2px 1px rgba(0, 0, 0, 0.8));
